Rename the add folder and correct linters errors
