+++ jest.config.js
@@ -1,5 +1,3 @@
 module.exports = {
-    testEnvironment: 'jsdom',
-    setupFiles: ['./src/__mocks__/mockLocalStore.js'],
-  };
-  
+  testEnvironment: 'jsdom',
+};
