--- src/__mocks__/mockLocalStore.js
@@ -1,20 +1,0 @@
-function mockStore () {
-    let dataStore = {};
-
-    return {
-      setItem: (key, value) => {
-        dataStore[key] = String(value) || '';
-      },
-      getItem: (key) => {
-        return dataStore[key] || null;
-      },
-      removeItem: (key) => {
-        delete dataStore[key];
-      },
-      get length() {
-        return Object.keys(dataStore).length;
-      },
-    };
-  }
-
-  export default mockStore;
